simple_ai_agents/social_media_agent/pyproject.toml
[project]
name = "social-media-agent"
version = "0.1.0"
description = "AI-powered social media automation tool that analyzes your Twitter writing style and generates new tweets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.48.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "composio>=0.8.9",
    "langchain-scrapegraph>=1.0.0",
    "langchain-nebius>=1.0.0",
    "langchain-core>=0.3.0",
    "memorisdk>=1.0.1",
    "pypdf>=5.1.0",
    "python-docx>=1.2.0",
]
